feat: verificando antes de excluir

+++ crud-angular/src/app/courses/containers/courses/courses.component.ts
@@ -7,6 +7,7 @@ import { catchError, Observable, of } from 'rxjs';
 import { ErrorDialogComponent } from '../../../shared/components/error-dialog/error-dialog.component';
 import { Course } from '../../model/course';
 import { CoursesService } from '../../services/courses.service';
+import { ConfirmationDialogComponent } from '../../../shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -55,19 +56,27 @@ export class CoursesComponent implements OnInit {
   }
 
   onDelete(course: Course) {
-    this.coursesService.deleteCouseById(course._id).subscribe(
-      () => {
-        this.refresh();
-        this._snackBar.open('Curso removido com sucesso! ✔️', '', {
-          duration: 5000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center',
-        });
-      },
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover esse curso?',
+    });
 
-      () => {
-        this.onError('Error ao tentar remover curso 🤥');
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.deleteCouseById(course._id).subscribe(
+          () => {
+            this.refresh();
+            this._snackBar.open('Curso removido com sucesso! ✔️', '', {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center',
+            });
+          },
+
+          () => {
+            this.onError('Error ao tentar remover curso 🤥');
+          }
+        );
       }
-    );
+    });
   }
 }
